feat(wallet-client): make bitcoin blockchain API used by client configurable
